// web/admin/src/request/types.ts
export enum NiuniuWikiContractDomainModelProvider {
  ModelProviderBrandBaiZhiCloud = "BaiZhiCloud",
  ModelProviderBaiZhiCloudModelStore = "BaiZhiCloudModelStore",
}
// ...
export enum DomainModelType {
  ModelTypeChat = "chat",
  ModelTypeEmbedding = "embedding",
  ModelTypeRerank = "rerank",
  ModelTypeAnalysis = "analysis",
  ModelTypeAnalysisVL = "analysis-vl",
}
// ...
export interface NiuniuWikiContractDomainModelListItem {
  api_header?: string;
  api_key?: string;
  /** for azure openai */
  api_version?: string;
  base_url?: string;
  completion_tokens?: number;
  id?: string;
  is_active?: boolean;
  model?: string;
  parameters?: NiuniuWikiContractDomainModelParam;
  prompt_tokens?: number;
  provider?: NiuniuWikiContractDomainModelProvider;
  total_tokens?: number;
  type?: DomainModelType;
}

export interface NiuniuWikiContractDomainModelParam {
  context_window?: number;
  max_tokens?: number;
  r1_enabled?: boolean;
  support_computer_use?: boolean;
  support_images?: boolean;
  support_prompt_cache?: boolean;
  temperature?: number;
}